[SLP][modularisation][NFC] Move reduction match helpers to SLPReductionUtils
Move the following BoUpSLP-independent reduction pattern-match helpers out
of SLPVectorizer.cpp into a new SLPVectorizer/SLPReductionUtils.{h,cpp}:

  getNonPhiOperand
  isReductionCandidate

matchRdxBop, used only by these two, moves alongside them and stays
file-local. Behavior is unchanged.

Part of the SLPVectorizer.cpp modularization effort:
https://discourse.llvm.org/t/modularizing-slpvectorizer-cpp/90922
